Silence compile warning.

+++ cycles/util/util_task.h
@@ -93,7 +93,7 @@ public:
 	static void exit();
 
 	/* number of threads that can work on task */
-	static int num_threads() { return threads.size(); }
+	static int num_threads() { return (int)threads.size(); }
 
 	/* test if any session is using the scheduler */
 	static bool active() { return users != 0; }
